Solved riddle for AoC 2024 day 25 part 1.

+++ aoc2024/aoc2024-kotlin/src/test/kotlin/de/havox_design/aoc2024/day25/CodeChronicleTest.kt
@@ -10,14 +10,14 @@ class CodeChronicleTest {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart1")
-    fun testProcessPart1(filename: String, expectedResult: Long) =
+    fun testProcessPart1(filename: String, expectedResult: Int) =
         CodeChronicle(filename).processPart1().shouldBe(expectedResult)
 
     companion object {
         @JvmStatic
         private fun getDataForTestProcessPart1(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2024/day25/day25sample.txt", 3L)
+                Arguments.of("de/havox_design/aoc2024/day25/day25sample.txt", 3)
             )
     }
 }
